requirements.txt
google-cloud-storage==1.23.0